requirements.txt
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Faker==27.0.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==3.0.3
importlib_metadata==8.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
mysql-connector-python==9.0.0
ordered-set==4.1.0
packaging==24.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rich==13.7.1
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.20.0